scarp/requirements.txt
PyQt6>=6.4.0
requests>=2.28.0
beautifulsoup4>=4.11.0
matplotlib>=3.6.0
python-docx>=0.8.11
selenium>=4.8.0
lxml>=4.9.0
